+++ src/__tests__/rjConstructor.test.js
@@ -1,5 +1,6 @@
 import rj from '../rj'
-import { isPartialRj, isObjectRj } from 'rocketjump-core'
+import { isObjectRj as strictIsObjectRj } from '../types'
+import { isPartialRj, isObjectRj, forgeRocketJump } from 'rocketjump-core'
 
 describe('rj constructor', () => {
   it('should produce rj object when called with effect or a function', () => {
@@ -57,7 +58,10 @@ describe('rj constructor', () => {
       /\[react-rocketjump\] effect should be defined only once, in the last argument/
     )
     expect(() => {
-      rj(() => 23, () => 777)
+      rj(
+        () => 23,
+        () => 777
+      )
     }).toThrowError(
       /\[react-rocketjump\] effect should be defined only once, in the last argument/
     )
@@ -118,11 +122,34 @@ describe('rj constructor', () => {
     }).toThrowError(/\[react-rocketjump\] you can't invoke a partialRj/)
 
     expect(() => {
-      rj(rj(() => 23), {
-        reducer: r => r,
-      })()
+      rj(
+        rj(() => 23),
+        {
+          reducer: r => r,
+        }
+      )()
     }).toThrowError(
       /\[react-rocketjump\] you can't pass an rj object as argument./
     )
+  })
+
+  it('should have strict isObjectRj', () => {
+    const dubRj = forgeRocketJump({
+      shouldRocketJump: () => true, // single invocation
+      makeRunConfig: () => null, // no run config
+      makeRecursionRjs: rjs => rjs, // don't touch configs
+      makeExport: (_, config, rjExport = {}) => {
+        return {
+          giova: 23,
+        }
+      },
+      finalizeExport: rjExport => ({ ...rjExport }), // don't hack config
+    })
+
+    expect(strictIsObjectRj(dubRj())).toBe(false)
+    expect(isObjectRj(dubRj())).toBe(true)
+
+    expect(strictIsObjectRj(rj(() => 2))).toBe(true)
+    expect(isObjectRj(rj(() => 2))).toBe(true)
   })
 })