+++ src/types.js
@@ -1,0 +1,6 @@
+import { isObjectRj as coreIsObjectRj } from 'rocketjump-core'
+import rj from './rj'
+
+export function isObjectRj(objRj) {
+  return coreIsObjectRj(objRj, rj)
+}