+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 2.2.0
+###### *March 18, 2019*
+
+No breaking changes only new features.
+
+#### `isObjectRj`
+
+Expose a new helper to check if a plain object is a **React** `RjObject`
+
+```js
+isObjectRj(RjObject) => Boolean
+```
+
+#### Use `deps` :gem: in action creators
+
+You can use `deps` in action creators:
+
+```js
+const [data, actions] = useRj(RjObject)
+actions.run(deps.maybe(false)) // Don't run
+actions.run(deps.maybeGet({ name: 'GioVa' }, 'name')) // Don't run as 'GioVa'
+actions.run(deps.withMeta(23, { id: 23 })) // Run with 23 and meta { id: 23 }
+```
+
+
 ## 2.1.1
 ###### *October 31, 2019*